# frontend/scripts/openapi/specs/auth-openapi.yaml
openapi: 3.1.0
info:
  title: Auth Service
  description: LazyMind authentication and authorization service (login, registration,
    token, user/role/group management)
  version: 1.0.0
paths:
...
  /api/authservice/v1/cloud/{provider}/connections:
    post:
      tags:
      - cloud-oauth
      summary: Create Connection
      operationId: create_connection_api_authservice_v1_cloud__provider__connections_post
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudConnectionCreateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/{provider}/oauth/app-credentials:
    get:
      tags:
      - cloud-oauth
      summary: Get Oauth App Credentials
      operationId: get_oauth_app_credentials_api_authservice_v1_cloud__provider__oauth_app_credentials_get
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudOAuthAppCredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - cloud-oauth
      summary: Save Oauth App Credentials
      operationId: save_oauth_app_credentials_api_authservice_v1_cloud__provider__oauth_app_credentials_put
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudOAuthAppCredentialBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudOAuthAppCredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - cloud-oauth
      summary: Delete Oauth App Credentials
      operationId: delete_oauth_app_credentials_api_authservice_v1_cloud__provider__oauth_app_credentials_delete
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudOAuthAppCredentialResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/{provider}/oauth/authorize-url:
    post:
      tags:
      - cloud-oauth
      summary: Oauth Authorize Url
      operationId: oauth_authorize_url_api_authservice_v1_cloud__provider__oauth_authorize_url_post
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudOAuthAuthorizeURLBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudOAuthAuthorizeURLResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/{provider}/oauth/callback:
    post:
      tags:
      - cloud-oauth
      summary: Oauth Callback
      operationId: oauth_callback_api_authservice_v1_cloud__provider__oauth_callback_post
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
          title: Provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudOAuthCallbackBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudOAuthCallbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections:
    get:
      tags:
      - cloud-oauth
      summary: List Connections
      operationId: list_connections_api_authservice_v1_cloud_connections_get
      security:
      - HTTPBearer: []
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      - name: auth_mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Auth Mode
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/status:batch:
    post:
      tags:
      - cloud-oauth
      summary: Batch Connection Status
      operationId: batch_connection_status_api_authservice_v1_cloud_connections_status_batch_post
      parameters:
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudConnectionStatusBatchBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionStatusBatchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/{connection_id}:
    get:
      tags:
      - cloud-oauth
      summary: Get Connection
      operationId: get_connection_api_authservice_v1_cloud_connections__connection_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - cloud-oauth
      summary: Delete Connection
      operationId: delete_connection_api_authservice_v1_cloud_connections__connection_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionDeleteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - cloud-oauth
      summary: Update Connection
      operationId: update_connection_api_authservice_v1_cloud_connections__connection_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudConnectionUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - cloud-oauth
      summary: Patch Connection
      operationId: patch_connection_api_authservice_v1_cloud_connections__connection_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CloudConnectionUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/{connection_id}/token:refresh:
    post:
      tags:
      - cloud-oauth
      summary: Refresh Connection Token
      operationId: refresh_connection_token_api_authservice_v1_cloud_connections__connection_id__token_refresh_post
      security:
      - HTTPBearer: []
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/internal/chat-enabled:
    get:
      tags:
      - cloud-oauth
      summary: List Chat Enabled Connections
      description: 'Internal endpoint: list connections with chat_enabled=true for
        a given owner.'
      operationId: list_chat_enabled_connections_api_authservice_v1_cloud_connections_internal_chat_enabled_get
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
      - name: owner_user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/internal/target-cache-candidates:
    get:
      tags:
      - cloud-oauth
      summary: List Target Cache Connections
      description: 'Internal endpoint: list active cloud connections for scan target
        cache prewarm.'
      operationId: list_target_cache_connections_api_authservice_v1_cloud_connections_internal_target_cache_candidates_get
      parameters:
      - name: provider
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: feishu
          title: Provider
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/{connection_id}/token:
    get:
      tags:
      - cloud-oauth
      summary: Get Connection Token
      operationId: get_connection_token_api_authservice_v1_cloud_connections__connection_id__token_get
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/authservice/v1/cloud/connections/{connection_id}/verify:
    get:
      tags:
      - cloud-oauth
      summary: Verify Connection
      operationId: verify_connection_api_authservice_v1_cloud_connections__connection_id__verify_get
      parameters:
      - name: connection_id
        in: path
        required: true
        schema:
          type: string
          title: Connection Id
      - name: user_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - name: tenant_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tenant Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CloudConnectionVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CloudConnectionCreateBody:
      properties:
        tenant_id:
          type: string
          title: Tenant Id
          default: ''
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        auth_mode:
          type: string
          title: Auth Mode
          default: tenant
        client_id:
          type: string
          title: Client Id
        client_secret:
          type: string
          title: Client Secret
        display_name:
          type: string
          title: Display Name
          default: ''
        provider_options:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Options
      type: object
      required:
      - client_id
      - client_secret
      title: CloudConnectionCreateBody
    CloudConnectionCreateResponse:
      properties:
        connection_id:
          type: string
          title: Connection Id
        tenant_id:
          type: string
          title: Tenant Id
        owner_user_id:
          type: string
          title: Owner User Id
          default: ''
        provider:
          type: string
          title: Provider
        auth_mode:
          type: string
          title: Auth Mode
        scope:
          type: string
          title: Scope
          default: ''
        status:
          type: string
          title: Status
          default: ACTIVE
      type: object
      required:
      - connection_id
      - tenant_id
      - provider
      - auth_mode
      title: CloudConnectionCreateResponse
    CloudConnectionDeleteResponse:
      properties:
        connection_id:
          type: string
          title: Connection Id
        status:
          type: string
          title: Status
          default: REVOKED
        deleted:
          type: boolean
          title: Deleted
          default: true
      type: object
      required:
      - connection_id
      title: CloudConnectionDeleteResponse
    CloudConnectionListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CloudConnectionResponse'
          type: array
          title: Items
      type: object
      required:
      - items
      title: CloudConnectionListResponse
...
    CloudConnectionStatusBatchBody:
      properties:
        connection_ids:
          items:
            type: string
          type: array
          maxItems: 500
          title: Connection Ids
      type: object
      title: CloudConnectionStatusBatchBody
    CloudConnectionStatusBatchResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/CloudConnectionStatusItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: CloudConnectionStatusBatchResponse
...
    CloudConnectionVerifyResponse:
      properties:
        connection_id:
          type: string
          title: Connection Id
        tenant_id:
          type: string
          title: Tenant Id
        owner_user_id:
          type: string
          title: Owner User Id
          default: ''
        provider:
          type: string
          title: Provider
        status:
          type: string
          title: Status
          default: ACTIVE
      type: object
      required:
      - connection_id
      - tenant_id
      - provider
      title: CloudConnectionVerifyResponse
    CloudOAuthAppCredentialBody:
      properties:
        client_id:
          type: string
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Secret
        provider_options:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Options
      type: object
      required:
      - client_id
      title: CloudOAuthAppCredentialBody
    CloudOAuthAppCredentialResponse:
      properties:
        provider:
          type: string
          title: Provider
        app_id:
          type: string
          title: App Id
          default: ''
        secret_configured:
          type: boolean
          title: Secret Configured
          default: false
        status:
          type: string
          title: Status
          default: ''
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
      type: object
      required:
      - provider
      title: CloudOAuthAppCredentialResponse
    CloudOAuthAuthorizeURLBody:
      properties:
        tenant_id:
          type: string
          title: Tenant Id
          default: ''
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        auth_mode:
          type: string
          title: Auth Mode
          default: oauth_user
        client_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Secret
        redirect_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Redirect Uri
        scope:
          anyOf:
          - type: string
          - type: 'null'
          title: Scope
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        reauthorize_connection_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reauthorize Connection Id
        provider_options:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Options
      type: object
      title: CloudOAuthAuthorizeURLBody
    CloudOAuthAuthorizeURLResponse:
      properties:
        connection_id:
          type: string
          title: Connection Id
        tenant_id:
          type: string
          title: Tenant Id
        owner_user_id:
          type: string
          title: Owner User Id
          default: ''
        provider:
          type: string
          title: Provider
        auth_mode:
          type: string
          title: Auth Mode
        scope:
          type: string
          title: Scope
          default: ''
        authorize_url:
          type: string
          title: Authorize Url
        state:
          type: string
          title: State
      type: object
      required:
      - connection_id
      - tenant_id
      - provider
      - auth_mode
      - authorize_url
      - state
      title: CloudOAuthAuthorizeURLResponse
    CloudOAuthCallbackBody:
      properties:
        tenant_id:
          type: string
          title: Tenant Id
          default: ''
        owner_user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner User Id
        connection_id:
          type: string
          title: Connection Id
        code:
          type: string
          title: Code
        state:
          anyOf:
          - type: string
          - type: 'null'
          title: State
        redirect_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Redirect Uri
      type: object
      required:
      - connection_id
      - code
      title: CloudOAuthCallbackBody
    CloudOAuthCallbackResponse:
      properties:
        connection_id:
          type: string
          title: Connection Id
        tenant_id:
          type: string
          title: Tenant Id
        owner_user_id:
          type: string
          title: Owner User Id
          default: ''
        provider:
          type: string
          title: Provider
        auth_mode:
          type: string
          title: Auth Mode
          default: oauth_user
        provider_account_id:
          type: string
          title: Provider Account Id
          default: ''
        display_name:
          type: string
          title: Display Name
          default: ''
        provider_tenant_key:
          type: string
          title: Provider Tenant Key
          default: ''
        provider_account_meta:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Provider Account Meta
        scope:
          type: string
          title: Scope
          default: ''
        status:
          type: string
          title: Status
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        refresh_token_bound:
          type: boolean
          title: Refresh Token Bound
          default: false
      type: object
      required:
      - connection_id
      - tenant_id
      - provider
      - status
      title: CloudOAuthCallbackResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer